Loading Graphs globally, rendering resource
Now we load graphs globally instead for a single resource. (Loading
graphs just for a single resource might be reimplemented at a later
point.) Instead of rendering the subject of a resource we now render
the resource itself. This requires to add another lens which renders
dcterms:subject with sublenses in order to get the old effect.

This breaks current configurations.