Workaround: rdflib_sparql skews order of variables
Usually one does:
for row = g.query("SELECT ?a ?b WHERE { ... }"):
    a = row[0]
    b = row[1]

This does not work. Since rdflib_sparql stores the variables as a set
instead of a list, the order of the tuple row is undefined. A
workaround is:

for row = g.query("SELECT ?a ?b WHERE { ... }").bindings:
    a = row[Variable("?a")]
    b = row[Variable("?b")]

This commit does not implement the workaround for every query!