Uploading tools from GIT to FTP
After publishing into a local git repository which is a copy of a
server instead of publishing directly to the server, one needs to
upload the changes via FTP.

image-diff creates a list of changes between to revisions in form of a
sequence of (abstract) FTP commands.

image-upload takes a list from image-diff, asks the user for passwords
for the FTP servers and executes the commands.