Updating even when store already exists
Running the updates is necessary even if the store contains already a
confGraph, because since last time, it could be that new triples have
been added and therefore the graph has to be updated.

Maybe I should split loading, updating and building entirely, or the
updates should be considered more like reasoning.