First version of the planner
This is just an experiment. It works quite well. Variants and
.htaccess generation are missing. The planner expects an n3 document
as input and writes a shell script to standard output.

In order to create the input, I use W3C's cwm with a plethora of
rules. This will allow me to reduce SemPipe's codebase drastically.
Documentation, examples and tests will be written later on.